+++ tests/testthat/test_part3_no4am.R
@@ -9,10 +9,10 @@ test_that("Part 3 runs fine on a short file not containing a full night", {
 
   # Create a test file that doesn't contain a 4am but does contain a noon and a midnight
   # (4am, noon and midnight all carry significance in the internal logic of g.sib.det)
-  create_test_acc_csv(Nmin=1000, starts_at_midnight = FALSE)
+  create_test_acc_csv(Nmin = 1000, starts_at_midnight = FALSE)
   GGIR(mode = c(1:3), datadir = fn, outputdir = getwd(), studyname = "test", f0 = 1, f1 = 1,
        do.report = c(), visualreport = FALSE, viewingwindow = 1,
-       do.parallel = FALSE, minimumFileSizeMB = 0)
+       do.parallel = FALSE, minimumFileSizeMB = 0, verbose = FALSE)
   
   out.file.path = paste0(dn, "/meta/ms3.out/", fn, ".RData")
   expect_true(file.exists(out.file.path))
@@ -21,10 +21,10 @@ test_that("Part 3 runs fine on a short file not containing a full night", {
 
   # Create a test file that doesn't contain a 4am or a noon, but does have a midnight
   # (4am, noon and midnight all carry significance in the internal logic of g.sib.det)
-  create_test_acc_csv(Nmin=720, start_time = "13:00:00")
+  create_test_acc_csv(Nmin = 720, start_time = "13:00:00")
   GGIR(mode = c(1:3), datadir = fn, outputdir = getwd(), studyname = "test", f0 = 1, f1 = 1,
        do.report = c(), visualreport = FALSE, viewingwindow = 1,
-       do.parallel = FALSE, minimumFileSizeMB = 0)
+       do.parallel = FALSE, minimumFileSizeMB = 0, verbose = FALSE)
   
   out.file.path = paste0(dn, "/meta/ms3.out/", fn, ".RData")
   expect_true(file.exists(out.file.path))