+++ NEWS.md
@@ -1,8 +1,11 @@
 # CHANGES IN GGIR VERSION 3.0-1
 
+- Part 2: Added possibility to csv file with date-oriented definition of study protocol #943
+
 - Part 5: New function to provide definition of the variable names in part 5 reports #938
 
 - Part 2: Bug fix in csv report generation when appending part 2 milestone data with differing number of columns #951
+
 
 # CHANGES IN GGIR VERSION 3.0-0